[Doc] Update documents
