+++ README.md
@@ -4,7 +4,7 @@
     <img src="./docs/images/logo.png" align="middle" width = "500" />
   </p>
 
-  **基于飞桨框架开发的高性能遥感图像处理开发套件，端到端地完成从训练到部署的全流程遥感深度学习应用。**
+  **基于飞桨框架开发的高性能遥感图像处理开发套件，帮助您端到端地完成从训练到部署的全流程遥感深度学习应用。**
 
   <!-- [![version](https://img.shields.io/github/release/PaddlePaddle/PaddleRS.svg)](https://github.com/PaddlePaddle/PaddleRS/releases) -->
   [![license](https://img.shields.io/badge/license-Apache%202-blue.svg)](LICENSE)
@@ -19,7 +19,7 @@
 
 ## 简介
 
-PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处理平台，支持遥感图像分类，目标检测，图像分割，以及变化检测等常用遥感任务，帮助开发者更便捷地完成从训练到部署全流程遥感深度学习应用。
+PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处理平台，支持遥感图像分类，目标检测，图像分割，以及变化检测等常用遥感任务，能够帮助遥感领域科研从业者和致力于遥感影像处理算法落地的开发者便捷地完成从训练到部署的全流程遥感深度学习应用。
 
 <div align="center">
 <img src="docs/images/whole_image.jpg"  width = "2000" />  
@@ -27,11 +27,17 @@ PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处
 
 ## 特性 <img src="./docs/images/feature.png" width="30"/>
 
-* <img src="./docs/images/f1.png" width="20"/> **特有的遥感数据处理模块**：针对遥感行业数据特点，提供了大尺幅数据切片与拼接，支持读取`tif`、`png`、 `jpeg`、 `bmp`、 `img`以及 `npy`等格式，支持地理信息保存和超分辨率。
+PaddleRS具有以下五大特色：
 
-* <img src="./docs/images/f2.png" width="20"/> **覆盖任务广**：支持目标检测、图像分割、变化检测、参数反演等多种任务
+* <img src="./docs/images/f1.png" width="20"/> **丰富的遥感特色模型**：支持FarSeg、BIT、ChangeStar等数十种遥感专用模型，覆盖图像分割、目标检测、变化检测等领域。
 
-* <img src="./docs/images/f3.png" width="20"/> **高性能**：支持多进程异步I/O、多卡并行训练、评估等加速策略，结合飞桨核心框架的显存优化功能，可大幅度减少分割模型的训练开销，让开发者更低成本、更高效地完成图像遥感图像的开发和训练。
+* <img src="./docs/images/f1.png" width="20"/> **面向遥感影像大幅面性质的针对性优化**：支持大幅面影像滑窗推理，使用内存延迟载入技术提升性能；支持对大幅面影像地理坐标信息的读写。
+
+* <img src="./docs/images/f2.png" width="20"/> **顾及遥感特性与地学知识的数据预处理**：针对遥感数据特点，提供对包含任意数量波段的数据以及多时相数据的预处理功能，支持影像配准、辐射校正、波段选择等遥感专有数据预处理方法。
+
+* <img src="./docs/images/f2.png" width="20"/> **对遥感专有任务的支持**：支持包括变化检测在内的遥感领域特色任务，提供完善的训练、部署教程以及丰富的实践案例。
+
+* <img src="./docs/images/f3.png" width="20"/> **工业级训练与部署性能**：支持多进程异步I/O、多卡并行训练等加速策略，结合飞桨核心框架的显存优化功能，可大幅度减少模型的训练开销，帮助开发者以更低成本、更高效地完成遥感的开发和训练。
 
 ## 产品矩阵<img src="./docs/images/model.png" width="30"/>
 
@@ -71,7 +77,7 @@ PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处
           <li>Faster R-CNN</li>
           <li>YOLOv3</li>
         </ul>
-        <b>超分/去噪</b><br>
+        <b>图像复原</b><br>
         <ul>
           <li>DRNet</li>
           <li>LESRCNNet</li>
@@ -119,9 +125,9 @@ PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处
         </ul>
         <b>数据预处理</b><br>
         <ul>
-          <li>data split</li>
-          <li>images match</li>
-          <li>bands select</li>
+          <li>影像切片</li>
+          <li>影像配准</li>
+          <li>波段选择</li>
         </ul>
       </td>
       <td>
@@ -150,31 +156,28 @@ PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处
   </tbody>
 </table>
 
-
 ### 代码结构
 
-这部分将展示PaddleRS的文件结构全貌。文件树如下：
+PaddleRS目录树中关键部分如下：
 
 ```
-├── deploy               # 部署相关的文档和脚本
-├── docs                 # 整个项目文档及图片
+├── deploy               # 部署相关文档与脚本
+├── docs                 # 项目文档
 ├── paddlers  
-│     ├── rs_models      # 遥感专用网络模型代码
-│     ├── datasets       # 数据加载相关代码
-│     ├── models         # 套件网络模型代码
-│     ├── tasks          # 相关任务代码
-│     ├── tools          # 相关脚本
-│     ├── transforms     # 数据处理及增强相关代码
-│     └── utils          # 各种实用程序文件
-├── tools                # 用于处理遥感数据的脚本
+│     ├── rs_models      # 遥感专用模型实现
+│     ├── datasets       # 数据集接口实现
+│     ├── models         # 视觉模型实现
+│     ├── tasks          # 训练器实现
+│     └── transforms     # 数据预处理/数据增强实现
+├── tools                # 遥感影像处理工具集
 └── tutorials
-      └── train          # 训练教程
+      └── train          # 模型训练教程
 ```
 
 ## 技术交流 <img src="./docs/images/chat.png" width="30"/>
 
-* 如果你发现任何PaddleRS存在的问题或者是建议, 欢迎通过[GitHub Issues](https://github.com/PaddlePaddle/PaddleRS/issues)给我们提issues。
-* 欢迎加入PaddleRS 微信群
+* 如果您发现任何PaddleRS存在的问题或是对PaddleRS有建议, 欢迎通过[GitHub Issues](https://github.com/PaddlePaddle/PaddleRS/issues)向我们提出。
+* 欢迎加入PaddleRS微信群
 <div align="center">
 <img src="./docs/images/wechat.png"  width = "150" />  
 </div>
@@ -183,38 +186,39 @@ PaddleRS是遥感科研院所、相关高校共同基于飞桨开发的遥感处
 
 * [快速上手PaddleRS](./tutorials/train/README.md)
 * 准备数据集
-   * [遥感数据介绍](./docs/data/rs_data_cn.md)
-   * [遥感数据集](./docs/data/dataset_cn.md)
-   * [智能标注工具EISeg](https://github.com/PaddlePaddle/PaddleSeg/tree/release/2.4/EISeg)
-   * [遥感数据处理脚本](./docs/data/tools.md)
+  * [遥感数据介绍](./docs/data/rs_data.md)
+  * [遥感数据集整理](./docs/data/dataset.md)
+  * [智能标注工具EISeg](https://github.com/PaddlePaddle/PaddleSeg/tree/release/2.4/EISeg)
+  * [遥感影像处理工具集](./docs/data/tools.md)
+* 组件介绍
+  * [数据预处理/数据增强](./docs/intro/transforms.md)
+  * [模型库](./docs/intro/model_zoo.md)
 * 模型训练
-   * [数据增强](./docs/apis/transforms.md)
-   * [模型库](./docs/apis/model_zoo.md)
-   * [模型训练说明](./docs/apis/model_zoo.md)
-   * 模型验证
-* 推理部署
-   * 模型导出
-   * 推理预测
-* 应用案例
-  * [变化检测示例](./docs/cases/csc_cd_cn.md)
-  * [超分模块示例](./docs/cases/sr_seg_cn.md)
+  * [模型训练API说明](./docs/apis/train.md)
+* 模型部署
+  * [模型导出](./deploy/export/README.md)
+  * [模型推理API说明](./docs/apis/infer.md)
+* 实践案例
+  * [遥感影像变化检测案例](./docs/cases/csc_cd_cn.md)
+  * [遥感影像超分辨率重建案例](./docs/cases/sr_seg_cn.md)
 * 代码贡献
-  * [PaddleRS代码注释规范](https://github.com/PaddlePaddle/PaddleRS/wiki/PaddleRS代码注释规范)
-
+  * [贡献指南](./docs/CONTRIBUTING.md)
+  * [开发指南](./docs/dev/dev_guide.md)
+  * [代码注释规范](./docs/dev/docstring.md)
 
 ## 开源贡献 <img src="./docs/images/love.png" width="30"/>
 
 * 非常感谢国家对地观测科学数据中心、中国科学院空天信息创新研究院、北京航空航天大学、武汉大学、中国石油大学（华东）、中国地质大学、中国四维、航天宏图、中科星图、超图等单位对PaddleRS项目的贡献。注：排名不分先后。
-* 非常感谢[geoyee](https://github.com/geoyee)(陈奕州), [Bobholamovic](https://github.com/Bobholamovic)(林漫晖), [kongdebug](https://github.com/kongdebug)(孔远杭), [huilin16](https://github.com/huilin16)(赵慧琳)等开发者对PaddleRS项目的贡献。
-* PaddleRS欢迎来自开源社区的贡献。如果您想要为PaddleRS贡献源码/案例，请参考[贡献指南](docs/CONTRIBUTING.md)。
+* 非常感谢[geoyee](https://github.com/geoyee)(陈奕州), [kongdebug](https://github.com/kongdebug)(孔远杭), [huilin16](https://github.com/huilin16)(赵慧琳)等开发者对PaddleRS项目的贡献。
+* PaddleRS欢迎来自开源社区的贡献。如果您想要为PaddleRS贡献源码/案例，请参考[贡献指南](./docs/CONTRIBUTING.md)。
 
 ## 许可证书
 
-本项目的发布受Apache 2.0 license许可认证。
+本项目的发布受[Apache 2.0 license](./LICENSE)许可认证。
 
 ## 学术引用 <img src="./docs/images/yinyong.png" width="30"/>
 
-如果我们的项目在学术上帮助到你，请考虑以下引用：
+如果我们的项目在学术上帮助到您，请考虑以下引用：
 
 ```latex
 @misc{paddlers2022,
