+++ docs/data/dataset.md
@@ -1,0 +1,19 @@
+# 遥感开源数据集
+
+PaddleRS搜集汇总了遥感领域常用的**开源**深度学习数据集，提供每个数据集的以下信息：数据集说明，影像信息，标注信息，源地址，AI Studio备份链接。这些数据集按任务类型可分为**图像分类、图像分割、变化检测、目标检测、目标跟踪、多标签分类、图像生成**等多种类型。目前已收录的遥感数据集有：
+
+* 图像分类数据集32个；
+* 目标检测数据集40个；
+* 图像分割数据集70个；
+* 变化检测数据集28个；
+* 实例分割数据集7个；
+* 多标签分类数据集3个；
+* 目标跟踪数据集9个；
+* 图像标题数据集3个；
+* 图像生成数据集8个。
+
+欢迎访问[遥感数据集汇总](./dataset_summary.md)，以获取详细信息。
+
+## 特别贡献数据集
+
+* 2020年中国典型城市道路样本数据CHN6-CUG，由中国地质大学[朱祺琪](http://grzy.cug.edu.cn/zhuqiqi)教授提供。相关介绍和下载方式请参考[此处](http://grzy.cug.edu.cn/zhuqiqi/zh_CN/yjgk/32368/content/1733.htm)。
