+++ docs/intro/model_zoo.md
@@ -1,0 +1,34 @@
+# 模型库
+
+PaddleRS的基础模型库来自Paddle-CV系列套件：[PaddleClas](https://github.com/PaddlePaddle/PaddleClas/blob/release/2.3/docs/zh_CN/algorithm_introduction/ImageNet_models.md)、[PaddleSeg](https://github.com/PaddlePaddle/PaddleSeg/blob/release/2.4/docs/model_zoo_overview_cn.md)、[PaddleDetection](https://github.com/PaddlePaddle/PaddleDetection/blob/release/2.3/README_cn.md#模型库)以及[PaddleGAN](https://github.com/PaddlePaddle/PaddleGAN/blob/develop/README_cn.md#模型库)。除此之外，PaddleRS也包含一系列遥感特色模型，可用于遥感影像分割、变化检测等。
+
+## PaddleRS支持的模型列表
+
+PaddleRS支持的全部模型如下（标注\*的为遥感专用模型）：
+
+| 任务 | 模型 |
+|--------|---------|
+| 变化检测 | \*BIT |
+| 变化检测 | \*CDNet |
+| 变化检测 | \*DSAMNet |
+| 变化检测 | \*DSIFN |
+| 变化检测 | \*SNUNet |
+| 变化检测 | \*STANet |
+| 变化检测 | \*FC-EF |
+| 变化检测 | \*FC-Siam-conc |
+| 变化检测 | \*FC-Siam-diff |
+| 场景分类 | HRNet |
+| 场景分类 | MobileNetV3 |
+| 场景分类 | ResNet50-vd |
+| 场景分类 | CondenseNetV2 |
+| 图像复原 | DRN |
+| 图像复原 | ESRGAN |
+| 图像复原 | LESRCNN |
+| 目标检测 | Faster R-CNN |
+| 目标检测 | PP-YOLO |
+| 目标检测 | PP-YOLO Tiny |
+| 目标检测 | PP-YOLOv2 |
+| 目标检测 | YOLOv3 |
+| 图像分割 | DeepLab V3+ |
+| 图像分割 | UNet |
+| 图像分割 | \*FarSeg |
